+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [v2.2.0] - 2019-12-01
+
 ### Added
 
 *   Future for SQL queries.
@@ -118,7 +120,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 *   Common domain exceptions for implementing a repository.
 
 
-[Unreleased]: https://github.com/falkoschumann/java-muspellheim-commons/compare/v2.1.0...HEAD
+[Unreleased]: https://github.com/falkoschumann/java-muspellheim-commons/compare/v2.2.0...HEAD
+[v221.0]: https://github.com/falkoschumann/java-muspellheim-commons/compare/v2.1.0...v2.2.0
 [v2.1.0]: https://github.com/falkoschumann/java-muspellheim-commons/compare/v2.0.0...v2.1.0
 [v2.0.0]: https://github.com/falkoschumann/java-muspellheim-commons/compare/v1.9.0...v2.0.0
 [v1.9.0]: https://github.com/falkoschumann/java-muspellheim-commons/compare/v1.8.0...v1.9.0